+++ Sources/MapboxCoreNavigation/Navigator.swift
@@ -4,7 +4,7 @@ import MapboxNavigationNative
 extension MapboxNavigationNative.Navigator {
     
     func status(at timestamp: Date) -> NavigationStatus {
-        return try! getStatusForMonotonicTimestampNanoseconds(
+        return getStatusForMonotonicTimestampNanoseconds(
             Int64(timestamp.nanosecondsSince1970)
         )
     }