+++ Sources/MapboxNavigation/Double.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension Double {
-    
-    func round(_ fraction: Int) -> Double {
-        let multiplier = pow(10, Double(fraction))
-        return Darwin.round(self * multiplier) / multiplier
-    }
-}