+++ Sources/TestHelper/TestHelper.h
@@ -6,6 +6,5 @@ FOUNDATION_EXPORT double TestHelperVersionNumber;
 //! Project version string for TestHelper.
 FOUNDATION_EXPORT const unsigned char TestHelperVersionString[];
 
-// In this header, you should import all the public headers of your framework using statements like #import <TestHelper/PublicHeader.h>
-
-#import <TestHelper/MMEEventsManager+Spy.h>
+#import "MMEEventsManager+Spy.h"
+#import "CLHeadingPrivate.h"