+++ Tests/MapboxCoreNavigationTests/RouteControllerTests.swift
@@ -1,8 +1,8 @@
 import XCTest
 import Turf
 import MapboxDirections
+import CoreLocation
 @testable import MapboxCoreNavigation
-#if !SWIFT_PACKAGE
 import TestHelper
 
 class RouteControllerTests: XCTestCase {
@@ -65,4 +65,3 @@ extension RouteControllerTests: RouterDataSource {
         return NavigationLocationManager.self
     }
 }
-#endif