+++ Tests/MapboxCoreNavigationTests/SimulatedLocationManagerTests.swift
@@ -1,8 +1,8 @@
 import XCTest
 import MapboxDirections
 import Turf
+import CoreLocation
 @testable import MapboxCoreNavigation
-#if !SWIFT_PACKAGE
 import TestHelper
 
 class SimulatedLocationManagerTests: XCTestCase {
@@ -64,4 +64,3 @@ class SimulatedLocationManagerSpy: NSObject, CLLocationManagerDelegate {
         self.locations.append(contentsOf: locations)
     }
 }
-#endif